+++ .github/scripts/sideloading-replace.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-filePath=packages/fx-core/src/common/m365/serviceConstant.ts
+filePath=packages/fx-core/src/component/m365/serviceConstant.ts
 echo "Replace placeholders in $filePath"
 sed -i -e "s@{{SERVICE_ENDPOINT_PLACEHOLDER}}@$SIDELOADING_SERVICE_ENDPOINT@g" $filePath
 sed -i -e "s@{{SERVICE_SCOPE_PLACEHOLDER}}@$SIDELOADING_SERVICE_SCOPE@g" $filePath
