+++ .github/workflows/env-checker-ci-pr.yml
@@ -9,7 +9,7 @@ on:
     paths:
       - ".github/workflows/env-checker-ci-pr.yml"
       - ".github/env-checker"
-      - "packages/fx-core/src/common/deps-checker/**"
+      - "packages/fx-core/src/component/deps-checker/**"
       - "packages/fx-core/tests/common/deps-checker/**"
   push:
     branches:
@@ -19,7 +19,7 @@ on:
     paths:
       - ".github/workflows/env-checker-ci-pr.yml"
       - ".github/env-checker"
-      - "packages/fx-core/src/common/deps-checker/**"
+      - "packages/fx-core/src/component/deps-checker/**"
       - "packages/fx-core/tests/common/deps-checker/**"
   workflow_dispatch: # Manual trigger
