+++ packages/fx-core/.nycrc
@@ -6,7 +6,7 @@
   ],
   "extension": [".js",".ts"],
   "exclude": [
-    "src/common/deps-checker/**/*",
+    "src/component/deps-checker/**/*",
     "src/question/generator.ts"
   ],
   "reporter": [
